% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calc_function.R
\name{additive.ll}
\alias{additive.ll}
\title{Function to Calculate Additive Log Likelihood for a Logistic Regression Model}
\usage{
additive.ll(beta, t)
}
\arguments{
\item{beta}{Vector of logistic regression coefficients.}

\item{t}{A 2x3 table of joint probabilities of disease and genotype.  Rows = case vs. control, columns=genotype.}
}
\value{
The log likelihood.
}
\description{
Calculates the log likelihood for a given set of logistic regression coefficients under an additive genetic model.
}
\examples{
additive.ll(c(-0.3793525, -1.1395417), 
 rbind(c(0.2339079, 0.05665039, 0.009441731), 
 c(0.3285921, 0.31834961, 0.053058269)))

}
