% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_function.R
\name{as.numeric2}
\alias{as.numeric2}
\title{Function to convert to numeric with scientific notation contaning the "." character}
\usage{
as.numeric2(char)
}
\arguments{
\item{char}{string to be converted to numeric}
}
\value{
a number
}
\description{
convert to numeric with scientific notation contaning the "." character
}
