% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calc_function.R
\name{calc.like.linear}
\alias{calc.like.linear}
\title{Function to Calculate Log Likelihood for a Linear Regression Model}
\usage{
calc.like.linear(beta, m, es_ab, es_bb, sd_y_x_model, sd_y_x_truth, model)
}
\arguments{
\item{beta}{Vector of linear regression coefficients.}

\item{m}{Minor allele frequency.}

\item{sd_y_x_model}{The standard deviation of Y (the outcome) given X (predictors/genotype) under the test model.}

\item{sd_y_x_truth}{The standard deviation of Y given X (predictors/genotype) given genotype under the true model.}

\item{model}{The genetic model in the linear regression: "Dominant", "Additive", "Recessive", "2df" or "null"}

\item{es}{Vector of effect sizes with two elements, (mean AB - mean AA) and (mean BB - mean AA).}
}
\value{
The log likelihood.
}
\description{
Convenience function to calculate the log likeilhood of a specified model.
}
