% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_function.R
\name{effect_size_function}
\alias{effect_size_function}
\title{Effect size function}
\usage{
effect_size_function(N = NULL, Case.Rate = NULL, k = NULL,
  MAF = NULL, power = NULL, risk_allele = T, Alpha = 0.05,
  True.Model = "All", Test.Model = "All")
}
\arguments{
\item{N}{Vector of the desired sample size(s)}

\item{Case.Rate}{Vector of the proportion(s) of cases in the sample (cases/(cases + controls)). Either k or Case.Rate must be specified.}

\item{k}{Vector of the number of controls per case. Either k or Case.Rate must be specified.}

\item{MAF}{Vector of minor allele frequencies}

\item{power}{Vector of powers to detect}

\item{risk_allele}{Logical: If OR > 1, the allele is classified as a "risk allele"}

\item{Alpha}{the desired type 1 error rate(s)}

\item{True.Model}{A vector vector the true underlying genetic model(s): 'Dominant', 'Additive', 'Recessive' or 'All'}

\item{Test.Model}{A vector specifying the assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive' or 'All'}
}
\description{
Calculates the odds ratio for a given power, at a given sample size, N, with type 1 error rate, Alpha
}
\examples{
or <- effect_size_function(N=c(100,1000), Case.Rate=seq(from=0.2,to=0.3,by=0.1), k=NULL, MAF= seq(from=0.2, to = 0.3, by = 0.025), 
  power=c(0.5, 0.8), Alpha = c(0.05), risk_allele = T, True.Model = 'All', Test.Model = 'All')

}
