% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_function.R
\name{find.prob.rec}
\alias{find.prob.rec}
\title{Recessive probability finding function}
\usage{
find.prob.rec(x, P_AA, P_AB, P_BB, cr, like)
}
\arguments{
\item{x}{Probability of disease given AB or BB}

\item{P_AA}{Probability the allele is homozygous for the major allele}

\item{P_AB}{Probability the allele is heteroygous}

\item{P_BB}{Probability the allele is homozygous for the minor allele}

\item{cr}{proportion of cases in the sample (cases/(cases + controls)).}

\item{like}{Expected log likelihood}
}
\value{
: The "a" in the binomial function ax^2 + bx + c that arises in solution for the additive OR functions
}
\description{
Operates within add.or.function to find probability of disease in a recessive truth given AB or BB, additive test model
}
