% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/power_function_linear_environment_interaction_logistic_mle.R
\name{integrand_funct_case}
\alias{integrand_funct_case}
\title{Function to generate integrand for mle for cases}
\usage{
integrand_funct_case(x1, x2)
}
\arguments{
\item{x1}{"true" part of model}

\item{x2}{"test" part of model}
}
\value{
a function to be used as the integrand for the mle
}
\description{
Returns the standard deviation of y given x for linear models with linear environment interaction
}
\examples{
integrand_funct_case(-1.462531 + 1*0.1823216,
	-1.462531 + 1*0.1823216) 

}
