% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_outcome_envir_interaction_mle_function.R
\name{linear.mles.log.envir.interaction}
\alias{linear.mles.log.envir.interaction}
\title{Function to calculate the standard deviation of y given x for linear models with logistic environment interaction}
\usage{
linear.mles.log.envir.interaction(MAF, P_e, ES_G, ES_E, ES_GE, Test.Model,
  True.Model, reduced = F)
}
\arguments{
\item{MAF}{Minor allele Frequency}

\item{P_e}{Population prevalence of logistic environmental factor}

\item{ES_G}{Genetic Effect size}

\item{ES_E}{Environment Effect size}

\item{ES_GE}{Environment x Genetic interaction Effect size}

\item{mod}{model}
}
\value{
The standard deviation of y given x for linear models with logistic environment interaction
}
\description{
Returns the standard deviation of y given x for linear models with logistic environment interaction
}
