% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mle_function.R
\name{linear.mles}
\alias{linear.mles}
\title{Function to calculate MLE's for linear models}
\usage{
linear.mles(m, es_ab, es_bb, model)
}
\arguments{
\item{m}{minor allele frequency}

\item{model}{The assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive', '2df'}

\item{es}{vector of effect sizes with two elements, (mean AB - mean AA) and (mean BB - mean AA)}
}
\value{
A vector of linear regression model coefficients.
}
\description{
Finds the maximum likelihood estiamtes for a given MAF under the specified genetic model and effect size.
}
