% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_outcome_envir_interaction_mle_function.R
\name{linear.outcome.log.envir.interaction.sds}
\alias{linear.outcome.log.envir.interaction.sds}
\title{Function to calculate the standard deviation of y given x for linear models with logistic environment interaction}
\usage{
linear.outcome.log.envir.interaction.sds(MAF, P_e, ES_G, ES_E, ES_GE, mod,
  True.Model, sd_y, reduced = F)
}
\arguments{
\item{MAF}{Minor allele Frequency}

\item{P_e}{Population prevalence of logistic environmental factor}

\item{ES_G}{Genetic Effect size}

\item{ES_E}{Environment Effect size}

\item{ES_GE}{Environment x Genetic interaction Effect size}

\item{mod}{Test model}

\item{True.Model}{True model}

\item{sd_y}{Standard deviation of y}

\item{reduced}{logical, indicates whether the X matrix will be used for a reduced model}
}
\value{
The standard deviation of y given x for linear models with logistic environment interaction
}
\description{
Returns the standard deviation of y given x for linear models with logistic environment interaction
}
