% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mle_function.R
\name{linear.sds}
\alias{linear.sds}
\title{Functions to Calculate Residual SD for Normal/Continuous Outcomes
Function to calculate the standard deviation of y given x for linear models}
\usage{
linear.sds(m, es_ab, es_bb, sd_y, model)
}
\arguments{
\item{m}{minor allele frequency}

\item{sd_y}{the standard deviation of y in the overall population.}

\item{model}{The assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive', '2df'}

\item{es}{vector of effect sizes with two elements, (mean AB - mean AA) and (mean BB - mean AA)}
}
\value{
A vector of linear regression model coefficients.
}
\description{
Functions to Calculate Residual SD for Normal/Continuous Outcomes
Function to calculate the standard deviation of y given x for linear models
}
