% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/power_function_environment_interaction_logistic_mle.R
\name{ll.ge.logistic}
\alias{ll.ge.logistic}
\title{Function to calculate MLE's for logistic models with logistic environment interaction}
\usage{
ll.ge.logistic(t, N = NULL, power = NULL, Alpha, mod,
  compareQuanto = 0)
}
\arguments{
\item{t}{A 2x6 table of the joint probabilities of disease, genotype, and environment. Rows are case vs. control and columns are genotypes.}

\item{N}{Sample size}

\item{power}{Power}

\item{Alpha}{Alpha}

\item{mod}{Test model}

\item{compareQuanto}{whether to include the factor to match with Quanto's calculations}
}
\value{
A vector of logistic regression model coefficients.
}
\description{
Finds the maximum likelihood estimates for a given 2x3 table under the specified genetic model.
}
