% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_mle_function.R
\name{logistic.mles}
\alias{logistic.mles}
\title{Function to calculate MLE's for logistic models}
\usage{
logistic.mles(t, model)
}
\arguments{
\item{t}{A 2x3 table of the joint probabilities of disease and genotype.  Rows are case vs. control and columns are genotypes.}

\item{model}{The assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive'}
}
\value{
A vector of logistic regression model coefficients.
}
\description{
Finds the maximum likelihood estimates for a given 2x3 table under the specified genetic model.
}
\examples{
logistic.mles(rbind(c(0.2967437, 0.1806723, 0.02258404), 
 c(0.3432563, 0.1393277, 0.01741596)), "Dominant")

}
