% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_function.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(x, min = 0, max = 1)
}
\arguments{
\item{x}{a number between 0 and 1.}

\item{min}{minimum}

\item{max}{maximum}
}
\value{
The logit of x.
}
\description{
Calculates the logit of a specified value.
}
\examples{
logit(0.5)

}
