% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_centrality_search_function.R
\name{ncp.search}
\alias{ncp.search}
\title{Function to Determine Non-Centrality Parameter of the Chi-squared distribution}
\usage{
ncp.search(x, power, stat, Alpha, df)
}
\arguments{
\item{x}{the non-centrality parameter}

\item{power}{the desired power}

\item{stat}{the detectable likelihood ratio test statistic}

\item{Alpha}{the desired type 1 error rate}

\item{df}{the degrees of freedom for the likelihood ratio test}
}
\value{
numeric value of the function
}
\description{
This function is set to 0 and solved for x, the non-centrality parameter to determine the sample size in \code{\link{ss.calc}}
}
