% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calc_function.R
\name{null.ll}
\alias{null.ll}
\title{Function to Calculate Null Log Likelihood for a Logistic Regression Model}
\usage{
null.ll(t)
}
\arguments{
\item{t}{A 2x3 table of joint probabilities of disease and genotype.  Rows = case vs. control, columns=genotype.}
}
\value{
The log likelihood.
}
\description{
Calculates the log likelihood for a given set of logistic regression coefficients under the null.
}
\examples{
null.ll(rbind(c(0.2339079, 0.05665039, 0.009441731), 
 c(0.3285921, 0.31834961, 0.053058269)))

}
