% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_calc.R
\name{or_calc}
\alias{or_calc}
\title{Odds ratio calculation}
\usage{
or_calc(a, b, c, d, e, f, mod, risk_allele)
}
\arguments{
\item{a}{The probability of a case given homogeneity for the major allele}

\item{b}{The probability of a case given heterozygosity}

\item{c}{The probability of a case given homogeneity for the minor allele}

\item{d}{The probability of a control given homogeneity for the major allele}

\item{e}{The probability of a control given heterozygosity}

\item{f}{The probability of a control given homogeneity for the minor allele#'}

\item{mod}{The model to be used (eg "Dominant", "Recessive", etc)}

\item{risk_allele}{Is this allele a risk allele? use T or F}
}
\value{
Odds ratio
}
\description{
Calculates odds ratio for given parameters. Used by the function odds_ratio_function.
}
\examples{
or_calc(a = 0.3649185, b = 0.12797197, c = 0.007109554, 
 d= 0.4450815, e= 0.05202803, f = 0.002890446, 
 mod = "Dominant", risk_allele = TRUE)

}
