% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_outcome_envir_interaction_mle_function.R
\name{p_vec_returner}
\alias{p_vec_returner}
\title{Function to output probability vector used in calculation of MLE's for linear outcome with logistic environment interaction}
\usage{
p_vec_returner(MAF, P_e)
}
\arguments{
\item{MAF}{Minor allele frequency}

\item{P_e}{Population prevalence of logistic environmental factor}
}
\value{
A probability vector to be used in MLE calculation for linear outcome with logistic environment interaction
}
\description{
Returns probability vector used in calculation of MLE's for linear outcome with logistic environment interaction
}
