% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/linear_outcome_linear_envir_interaction_mle_function.R
\name{p_vec_returner_lin_env}
\alias{p_vec_returner_lin_env}
\title{Function to output probability vector used in calculation of MLE's for linear outcome with linear environment interaction}
\usage{
p_vec_returner_lin_env(MAF)
}
\arguments{
\item{MAF}{Minor Allele Frequency}
}
\value{
A probability vector to be used in MLE calculation for linear outcome with linear environment interaction
}
\description{
Returns probability vector used in calculation of MLE's for linear outcome with linear environment interaction
}
\examples{
p_vec_returner_lin_env(0.1)

}
