% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_power_function_environment_interaction.R
\name{power_envir.calc.linear_outcome}
\alias{power_envir.calc.linear_outcome}
\title{Function to Calculate Power for Linear Models with logistic environment interaction}
\usage{
power_envir.calc.linear_outcome(N = NULL, MAF = NULL, ES_G = NULL,
  ES_E = NULL, ES_GE = NULL, P_e = NULL, R2_G = NULL,
  R2_E = NULL, R2_GE = NULL, sd_y = NULL, Alpha = 0.05,
  True.Model = "All", Test.Model = "All", compareQuanto = 0)
}
\arguments{
\item{N}{Vector of the desired sample size(s)}

\item{MAF}{Vector of minor allele frequencies}

\item{sd_y}{Standard deviation of the outcome in the population (ignoring genotype). Either sd_y_x or sd_y must be specified.}

\item{Alpha}{the desired type 1 error rate(s)}

\item{True.Model}{A vector specifying the true underlying genetic model(s): 'Dominant', 'Additive', 'Recessive' or 'All'}

\item{Test.Model}{A vector specifying the assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive' or 'All'}

\item{ES}{Vector of effect sizes (difference in means) to detect. Either ES or R2 must be specified.}

\item{R2}{Vector of R-squared values to detect. Either ES or R2 must be specified.}
}
\value{
A data frame including the power for all combinations of the specified parameters (Case.Rate, ES, Power, etc)
}
\description{
Calculates the power to detect an difference in means/effect size/regression coefficient, at a given sample size, N, with type 1 error rate, Alpha
}
\examples{
pw <- power.calc.linear(N=c(1000,2000),
    MAF=seq(0.05, 0.1, 0.01), ES=c(3,4),sd_y = c(1,2),Alpha=c(0.05),
    True.Model='All', Test.Model='All')

}
