% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_roots_function.R
\name{quad_roots}
\alias{quad_roots}
\title{Function to Solve Quadratic Equations}
\usage{
quad_roots(a, b, c)
}
\arguments{
\item{a}{the coefficient for x^2}

\item{b}{the coefficient for x}

\item{c}{the constant}
}
\value{
The positive root of the quadratic equation ax^2 + bx + c
}
\description{
Finds the positive root of a quadratic equation ax^2 + bx + c .
}
\examples{
pw<-power.calc(N=c(1000,2000), Case.Rate=c(0.5), 
	k=NULL, MAF=seq(0.05, 0.1, 0.01), OR=c(3,4),
	Alpha=c(0.05), True.Model='All', Test.Model='All')

}
