% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calc_function.R
\name{recessive.ll}
\alias{recessive.ll}
\title{Function to Calculate Recessive Log Likelihood for a Logistic Regression Model}
\usage{
recessive.ll(beta, t)
}
\arguments{
\item{beta}{Vector of logistic regression coefficients.}

\item{t}{A 2x3 table of joint probabilities of disease and genotype.  Rows = case vs. control, columns=genotype.}
}
\value{
The log likelihood.
}
\description{
Calculates the log likelihood for a given set of logistic regression coefficients under a recessive genetic model.
}
