% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_function.R
\name{solve_a}
\alias{solve_a}
\title{Binomial coefficient calculation}
\usage{
solve_a(b, cr, P_AA, P_AB, P_BB)
}
\arguments{
\item{b}{The "b" in the binomial function ax^2 + bx + c that arises in solution for the additive OR functions}

\item{cr}{proportion of cases in the sample (cases/(cases + controls)).}

\item{P_AA}{Probability the allele is homozygous for the major allele}

\item{P_AB}{Probability the allele is heteroygous}

\item{P_BB}{Probability the allele is homozygous for the minor allele}
}
\value{
: The "a" in the binomial function ax^2 + bx + c that arises in solution for the additive OR functions
}
\description{
Operates within add.or.function to solve for 'a' when 'b' is known in an additive model
}
