% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_outcome_envir_interaction_mle_function.R
\name{X_mat_returner}
\alias{X_mat_returner}
\title{Function to output X matrices used in calculation of MLE's for linear outcome with logistic environment interaction}
\usage{
X_mat_returner(mod, reduced = F)
}
\arguments{
\item{mod}{type of model}

\item{reduced}{logical, indicates whether the X matrix will be used for a reduced model}
}
\value{
A matrix to be used in MLE calculation for linear outcome with logistic environment interaction
}
\description{
Returns X matrices used in calculation of MLE's for linear outcome with logistic environment interaction
}
\examples{
X_mat_returner(mod = "Dominant")

}
