% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/linear_outcome_linear_envir_interaction_mle_function.R
\name{X_mat_returner_lle}
\alias{X_mat_returner_lle}
\title{Function to output X matrices used in calculation of MLE's for linear outcome with linear environment interaction}
\usage{
X_mat_returner_lle(mod)
}
\arguments{
\item{mod}{type of model}
}
\value{
A probability vector to be used in MLE calculation for linear outcome with linear environment interaction
}
\description{
Function to output X matrices used in calculation of MLE's for linear outcome with linear environment interaction
}
