% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_outcome_envir_interaction_mle_function.R
\name{calc.like.linear.log.envir.interaction}
\alias{calc.like.linear.log.envir.interaction}
\title{Function to calculate the standard deviation of y given x for linear models with logistic environment interaction}
\usage{
calc.like.linear.log.envir.interaction(beta_hat, MAF, P_e, ES_G, ES_E,
  ES_GE, sd_y_x_truth, sd_y_x_model, Test.Model, True.Model, reduced = F)
}
\arguments{
\item{beta_hat}{Effect sizes from MLE}

\item{MAF}{Minor allele Frequency}

\item{P_e}{Population prevalence of logistic environmental factor}

\item{ES_G}{Genetic Effect size}

\item{ES_E}{Environment Effect size}

\item{ES_GE}{Environment x Genetic interaction Effect size}

\item{sd_y_x_truth}{Standard deviation of y for the true model}

\item{True.Model}{True model}

\item{sd_y_x_test}{Standard deviation of y for the test model}

\item{sd_y}{Standard deviation of y}

\item{Test.model}{Test model}
}
\value{
The standard deviation of y given x for linear models with logistic environment interaction
}
\description{
Returns the standard deviation of y given x for linear models with logistic environment interaction
}
