% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_calc_function.R
\name{calc.like}
\alias{calc.like}
\title{Function to Calculate Log Likelihood for a Logistic Regression Model}
\usage{
calc.like(beta, t, model)
}
\arguments{
\item{beta}{Vector of logistic regression coefficients.}

\item{t}{A 2x3 table of joint probabilities of disease and genotype.  Rows = case vs. control, columns=genotype.}

\item{model}{The genetic model in the logisitic regression: "Dominant", "Additive", "Recessive", "2df" or "null"}
}
\value{
The log likelihood.
}
\description{
Convenience function to calculate the log likelihood of a specified model.
}
