% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/linear_outcome_linear_envir_interaction_mle_function.R
\name{expected.linear.ll.lin.env}
\alias{expected.linear.ll.lin.env}
\title{Function to Calculate Expected Log Likelihood for a Single Genotype with linear environment interaction}
\usage{
expected.linear.ll.lin.env(sd_y_x_model)
}
\arguments{
\item{sd_y_x_model}{The standard deviation of Y (the outcome) given X (predictors/genotype) under the test model.}
}
\value{
The log likelihood.
}
\description{
Calculates the expected log likelihood for a single genotype with linear environment interaction 
 given the true and estimated mean and standard deviation for the outcome.
}
\examples{
expected.linear.ll.lin.env(4.309354)

}
