% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_mle_function.R
\name{linear.sds}
\alias{linear.sds}
\title{Functions to Calculate Residual SD for Normal/Continuous Outcomes
Function to calculate the standard deviation of y given x for linear models}
\usage{
linear.sds(m, es_ab, es_bb, sd_y, model)
}
\arguments{
\item{m}{minor allele frequency}

\item{es_ab}{effect size for mean AB - mean AA}

\item{es_bb}{effect size for mean BB - mean AA}

\item{sd_y}{the standard deviation of y in the overall population.}

\item{model}{The assumed genetic model(s) used in testing: 'Dominant', 'Additive', 'Recessive', '2df'}
}
\value{
A vector of linear regression model coefficients.
}
\description{
Functions to Calculate Residual SD for Normal/Continuous Outcomes
Function to calculate the standard deviation of y given x for linear models
}
\examples{
linear.sds(m = 0.1, es_ab = 0, es_bb = 3, sd_y = 1, model = "Dominant")

}
