% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_function.R
\name{power.plot}
\alias{power.plot}
\title{Function to Plot Power Results}
\usage{
power.plot(data = NULL, x = "MAF", panel.by = "True.Model",
  select.Alpha = NULL, select.OR = NULL, select.N = NULL,
  select.MAF = NULL, select.Case.Rate = NULL, select.True.Model = NULL,
  select.Test.Model = NULL)
}
\arguments{
\item{data}{The data frame result from \code{\link{power.calc}}}

\item{x}{The desired variable on the y axis: "MAF", "OR", "Alpha", or "N_total"}

\item{panel.by}{A grouping variable to panel the graphs by: "True.Model", "MAF", "OR", "Alpha", or "N_total"}

\item{select.Alpha}{Only produce graphs for the specified Alpha level(s).}

\item{select.OR}{Only produce graphs for the specified odds ratio(s).}

\item{select.N}{Only produce graphs for the specified sample size(s).}

\item{select.MAF}{Only produce graphs for the specified minor allele frequency(ies).}

\item{select.Case.Rate}{Only produce graphs for the specified case rate(s).}

\item{select.True.Model}{Only produce graphs for the specified true genetic model(s): "Additive1", "Additive2", "Dominant", "Recessive".}

\item{select.Test.Model}{Only produce graphs for the specified testing model(s): "Additive", "Dominant", "Recessive", "2df".}
}
\value{
A series of plots with sample size on the Y axis.
}
\description{
Plot the power results by MAF, OR, Alpha or N
}
\examples{
pw<-power.calc(N=c(1000,2000), Case.Rate=c(0.5), k=NULL, 
    MAF=seq(0.05, 0.1, 0.01), OR=c(3,4),Alpha=c(0.05), 
    True.Model='All', Test.Model='All')
power.plot(data=pw, x='MAF')

}
